refactor(frontend): split task-detail-modal into sub-components
Extract TaskFormFields, TaskSubtaskSection, and TaskDeliverableSection
into app/components/task-detail/. The main modal now orchestrates layout
and data flow while sub-components own their rendering logic.

Closes #394
