+++ app/components/task-detail/index.ts
@@ -1,0 +1,4 @@
+export { TaskFormFields, initialForm } from "./task-form-fields";
+export type { TaskForm } from "./task-form-fields";
+export { TaskSubtaskSection } from "./task-subtask-section";
+export { TaskDeliverableSection } from "./task-deliverable-section";
